Archive_Etudiant/requirements.txt
numpy
keras
tensorflow==2.15.0
onnx
tf2onnx 
pillow
scikit-learn
matplotlib
IPython
Pandas
